// packages/frontend/src/api/models/ExerciseCreationPatientBasicInfo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ExerciseCreationPatientBasicInfo
 */
export interface ExerciseCreationPatientBasicInfo {
    /**
     * 
     * @type {string}
     * @memberof ExerciseCreationPatientBasicInfo
     */
    age: string;
    /**
     * 
     * @type {string}
     * @memberof ExerciseCreationPatientBasicInfo
     */
    sex: string;
    /**
     * 
     * @type {string}
     * @memberof ExerciseCreationPatientBasicInfo
     */
    weight: string;
    /**
     * 
     * @type {string}
     * @memberof ExerciseCreationPatientBasicInfo
     */
    height: string;
}

/**
 * Check if a given object implements the ExerciseCreationPatientBasicInfo interface.
 */
export function instanceOfExerciseCreationPatientBasicInfo(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "age" in value;
    isInstance = isInstance && "sex" in value;
    isInstance = isInstance && "weight" in value;
    isInstance = isInstance && "height" in value;

    return isInstance;
}

export function ExerciseCreationPatientBasicInfoFromJSON(json: any): ExerciseCreationPatientBasicInfo {
    return ExerciseCreationPatientBasicInfoFromJSONTyped(json, false);
}

export function ExerciseCreationPatientBasicInfoFromJSONTyped(json: any, ignoreDiscriminator: boolean): ExerciseCreationPatientBasicInfo {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'age': json['age'],
        'sex': json['sex'],
        'weight': json['weight'],
        'height': json['height'],
    };
}

export function ExerciseCreationPatientBasicInfoToJSON(value?: ExerciseCreationPatientBasicInfo | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'age': value.age,
        'sex': value.sex,
        'weight': value.weight,
        'height': value.height,
    };
}